Extract shared filter functions into video-filters module, fix overlay positioning
The buildVideoFilter, buildAudioFilter, buildAudioTrimFilter, and
buildArguments functions were copy-pasted between ffmpeg.ts and
ffmpeg.worker.ts, and the two copies had diverged. The worker copy
used W and H in overlay positioning expressions, but in FFmpeg's
overlay filter those refer to the overlay dimensions, not the main
video. This caused overlays at any position except top-left to render
at negative coordinates in exported videos.

Moving the shared logic into a single video-filters.ts module
eliminates the duplication and fixes the root cause. The overlay
position map now uses main_w and main_h throughout. ffmpeg.ts
re-exports buildVideoFilter and buildAudioFilter from the shared
module, preserving the public API for existing consumers.
